+++ docs/changelog.rst
@@ -20,6 +20,7 @@ Non-API breaking changes:
 * added predicate databasing
 * added name mangling to `PicklingDatabases`
 * fixed pickle to use highest protocol in `PicklingDatabases`
+* Pathpoint will try to read it's previous current value upon startup
 
 v0.13
 ~~~~~